Add test to shooter class and attempt fix on clnode class
The shooter objects can be destroyed. I think the issue is that the
CLNode objects recursively call the destructor each time they are called
since they also have elements that are their own time. This can get
complicated. I am going to remove the reference to the destructor and
see if that fixes anything.