Add first scenario to CLQueue tests
The first scenario was written and it works, all the tests are passing.

Something to keep in mind is that the queue is really just the
circularly linked list, so much of it doesn't have to be tested. Just
the methods that are more than calling a singular method from the CLList
class.