Add test for dequeuing from queue
Added a test/assertion to remove an element from the queue and make sure
it is working. I had to remove the generated number of shooters value 1
because then there are no elements left in the queue and it throws an
error, which is good!

No need to write a scenario where we test more of the dequeing since
dequeing works exactly the same as removing from a circular list.